remove interface
