+++ thain-server/src/main/java/com/xiaomi/thain/server/service/X5Service.java
@@ -5,54 +5,75 @@
  */
 package com.xiaomi.thain.server.service;
 
+import com.alibaba.fastjson.JSON;
+import com.xiaomi.thain.common.exception.ThainRuntimeException;
+import com.xiaomi.thain.server.dao.X5Dao;
 import com.xiaomi.thain.server.model.X5Config;
+import com.xiaomi.thain.server.model.dp.X5ConfigDp;
 import com.xiaomi.thain.core.model.dr.X5ConfigDr;
 import com.xiaomi.thain.server.model.rq.X5ConfigRq;
 import lombok.NonNull;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 /**
  * @author liangyongrui@xiaomi.com
- * @date 19-5-7 上午11:40
+ * @date 19-5-7 上午11:01
  */
-public interface X5Service {
-    /**
-     * get Config By AppId
-     *
-     * @param appid appId
-     * @return see {@link X5Config}
-     */
-    X5Config getX5Config(@NonNull String appid);
+@Slf4j
+@Service
+public class X5Service {
 
-    /**
-     * get All Configs
-     *
-     * @return see {@link X5ConfigDr}
-     */
-    List<X5ConfigDr> getAllConfigs();
+    @NonNull
+    private final X5Dao x5Dao;
 
-    /**
-     * delete Config
-     *
-     * @param appId AppId
-     */
-    void deleteX5Config(@NonNull String appId);
+    public X5Service(@NonNull X5Dao x5Dao) {
+        this.x5Dao = x5Dao;
+    }
 
-    /**
-     * insert Config
-     *
-     * @param x5ConfigRq see {@link X5ConfigRq}
-     * @return false-failure true-success
-     */
-    boolean insertX5Config(@NonNull X5ConfigRq x5ConfigRq);
+    public X5Config getX5Config(@NonNull String appId) {
+        return x5Dao.getX5Config(appId).orElseThrow(() -> new ThainRuntimeException("AppId is not Existed"));
+    }
 
-    /**
-     * update Config
-     *
-     * @param x5ConfigRq see {@link X5ConfigRq}
-     * @return false-failure true-success
-     */
-    boolean updateX5Config(@NonNull X5ConfigRq x5ConfigRq);
+    public List<X5ConfigDr> getAllConfigs() {
+        return x5Dao.getAllX5Config();
+    }
+
+    public void deleteX5Config(@NonNull String appId) {
+        x5Dao.deleteX5Config(appId);
+    }
+
+    public boolean insertX5Config(@NonNull X5ConfigRq x5ConfigRq) {
+        if (x5ConfigRq.principals.isEmpty()) {
+            return false;
+        }
+        if (!x5Dao.getX5Config(x5ConfigRq.appId).isPresent()) {
+            x5Dao.addX5Config(X5ConfigDp.builder()
+                    .appId(x5ConfigRq.appId)
+                    .appKey(x5ConfigRq.appKey)
+                    .appName(x5ConfigRq.appName)
+                    .description(x5ConfigRq.description)
+                    .principal(JSON.toJSONString(x5ConfigRq.principals))
+                    .build());
+            return true;
+        }
+        return false;
+    }
+
+    public boolean updateX5Config(@NonNull X5ConfigRq x5ConfigRq) {
+        if (x5Dao.getX5Config(x5ConfigRq.appId).isPresent()) {
+            x5Dao.updateX5Config(X5ConfigDp.builder()
+                    .appId(x5ConfigRq.appId)
+                    .appKey(x5ConfigRq.appKey)
+                    .appName(x5ConfigRq.appName)
+                    .description(x5ConfigRq.description)
+                    .principal(JSON.toJSONString(x5ConfigRq.principals))
+                    .build());
+            return true;
+        }
+        return false;
+    }
 }
